// client/src/js/bindings/HytaleDeviceFlowStartResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Response when starting a Hytale device flow
 */
export type HytaleDeviceFlowStartResponse = { session_id: string, user_code: string, verification_uri: string, verification_uri_complete: string, expires_in: number, interval: number, };